HYDRO-18: Adding base endpoints for grow chamber history data

+++ src/main/resources/db/migration/V1.4.29.1__Add_Grow_Chamber_History_Table.sql
@@ -13,15 +13,15 @@
 DROP TABLE grow_chamber_log;
 
 CREATE TABLE grow_chamber_history (
-  id INT(10) UNSIGNED NOT NULL AUTO_INCREMENT,
-  system_id INT(10) UNSIGNED NOT NULL,
-  ph FLOAT(2,2) UNSIGNED NOT NULL DEFAULT 0.0,
-  tds FLOAT(3,2) UNSIGNED NOT NULL DEFAULT 0.0,
-  water_temp_celsius FLOAT(3,2) UNSIGNED NOT NULL DEFAULT 0.0,
-  air_temp_celsius FLOAT(3,2) UNSIGNED NOT NULL DEFAULT 0.0,
-  humidity FLOAT(3,2) UNSIGNED NOT NULL DEFAULT 0.0,
-  light_status TINYINT(1) UNSIGNED NOT NULL DEFAULT 0,
-  insert_date_utc DATETIME NOT NULL DEFAULT current_timestamp(3),
+  id                               INT(10)      UNSIGNED NOT NULL AUTO_INCREMENT,
+  system_id                        INT(10)      UNSIGNED NOT NULL,
+  ph                               DECIMAL(5,2) UNSIGNED NOT NULL DEFAULT 0.0,
+  tds                              DECIMAL(5,2) UNSIGNED NOT NULL DEFAULT 0.0,
+  water_temp_celsius               DECIMAL(5,2) UNSIGNED NOT NULL DEFAULT 0.0,
+  air_temp_celsius                 DECIMAL(5,2) UNSIGNED NOT NULL DEFAULT 0.0,
+  humidity                         DECIMAL(5,2) UNSIGNED NOT NULL DEFAULT 0.0,
+  light_status                     TINYINT(1)   UNSIGNED NOT NULL DEFAULT 0,
+  insert_date_utc                  DATETIME              NOT NULL DEFAULT current_timestamp(3),
   PRIMARY KEY (id));
 
 -- ---------------------------------------------------------------------------------
